Agregar validación para la eliminación de sucursales, asegurando que no existan usuarios asociados

+++ src/sucursales/sucursales.service.ts
@@ -5,16 +5,19 @@ import {
 } from '@nestjs/common';
 import { CreateSucursalDto } from './dto/create-sucursal.dto';
 import { UpdateSucursalDto } from './dto/update-sucursal.dto';
+import { QuerySucursalDto } from './dto/query-sucursal.dto';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Usuario } from 'src/usuarios/entities/usuario.entity';
 import { Sucursal } from './entities/sucursal.entity';
 import { Repository } from 'typeorm';
-import { QuerySucursalDto } from './dto/query-sucursal.dto';
 
 @Injectable()
 export class SucursalesService {
   constructor(
     @InjectRepository(Sucursal)
     private sucursalesRepository: Repository<Sucursal>,
+    @InjectRepository(Usuario)
+    private usuariosRepository: Repository<Usuario>,
   ) {}
 
   async create(createSucursalDto: CreateSucursalDto): Promise<Sucursal> {
@@ -169,21 +172,24 @@ export class SucursalesService {
   }  
 
   async remove(id: number): Promise<{ message: string; sucursal?: Sucursal }> {
-    const sucursal = await this.findOne(id);
-  
-    const usuariosAsociados = await this.sucursalesRepository
-      .createQueryBuilder('sucursal')
-      .leftJoinAndSelect('sucursal.usuarios', 'usuario')
-      .where('sucursal.id = :id', { id })
-      .getCount();
-  
+    const sucursal = await this.sucursalesRepository.findOne({ where: { id } });
+
+    if (!sucursal) {
+      throw new NotFoundException(`La sucursal con ID ${id} no existe.`);
+    }
+
+    const usuariosAsociados = await this.usuariosRepository.count({
+      where: { sucursalId: id },
+    });
+
     if (usuariosAsociados > 0) {
       throw new BadRequestException(
         `La sucursal con ID ${id} no puede ser eliminada porque tiene usuarios asociados.`,
       );
     }
-  
+
     await this.sucursalesRepository.remove(sucursal);
+
     return {
       message: 'La sucursal ha sido eliminada exitosamente',
       sucursal,
